REF: Rename `tabs` module to `layouts`
